Make the volume mount readOnly
I found that there are multiple ways to mount a PVC readOnly. The
current method of doing it will mount the RWX PVC to a host as readOnly.
There are some CSI providers (well at least cephfs), that have problems
with that. So to maximize compatibility I move the readOnly flag from
the VolumeSource to the VolumeMount as this will mount the volumes
readWrite on the host and then bindmount it to the containers as
readOnly.